Add functionality for ship movement up and down the screen
The player cannot do much with a static ship. Responding to user input is necessary.

Create function update in Ship class to handle movement of the ship.

Update check_events function in game_functions module to handle relevant player inputs.

Add an attribute to Settings class to control the movement speed of the ship.

Update the main loop in sideways_shooter.py to update the ship's position.